requirements.txt
# Neural Feature Alignment Dependencies
# Tested with Python 3.9+

# Deep Learning
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# EEG Processing
mne>=1.0.0

# Signal Processing
PyWavelets>=1.1.0
antropy>=0.1.4

# Graph
networkx>=2.6.0

# Utils
joblib>=1.0.0
PyYAML>=6.0
tqdm>=4.60.0
